refactor(TermsAndConditions): migrate to functional component and update tests
