--- app/components/Views/TermsAndConditions/index.js
@@ -1,66 +1,0 @@
-import React, { PureComponent } from 'react';
-import { Text, StyleSheet, TouchableOpacity } from 'react-native';
-import PropTypes from 'prop-types';
-import { fontStyles } from '../../../styles/common';
-import { strings } from '../../../../locales/i18n';
-import AppConstants from '../../../core/AppConstants';
-import { ThemeContext, mockTheme } from '../../../util/theme';
-import { TermsAndConditionsSelectorsIDs } from './TermsAndConditions.testIds';
-
-const createStyles = (colors) =>
-  StyleSheet.create({
-    text: {
-      ...fontStyles.normal,
-      color: colors.text.alternative,
-      textAlign: 'center',
-      fontSize: 10,
-    },
-    link: {
-      textDecorationLine: 'underline',
-    },
-  });
-
-/**
- * View that is displayed in the flow to agree terms and conditions
- */
-export default class TermsAndConditions extends PureComponent {
-  static propTypes = {
-    /**
-    /* navigation object required to push and pop other views
-    */
-    navigation: PropTypes.object,
-  };
-
-  press = () => {
-    const { navigation } = this.props;
-    navigation.navigate('Webview', {
-      screen: 'SimpleWebview',
-      params: {
-        url: AppConstants.URLS.TERMS_AND_CONDITIONS,
-        title: strings('terms_and_conditions.title'),
-      },
-    });
-  };
-
-  render() {
-    const colors = this.context.colors || mockTheme.colors;
-    const styles = createStyles(colors);
-
-    return (
-      <TouchableOpacity
-        testID={TermsAndConditionsSelectorsIDs.ACCEPT_BUTTON}
-        onPress={this.press}
-      >
-        <Text style={styles.text}>
-          {strings('terms_and_conditions.description')}
-          <Text style={styles.link}>
-            {strings('terms_and_conditions.terms')}
-          </Text>
-          .
-        </Text>
-      </TouchableOpacity>
-    );
-  }
-}
-
-TermsAndConditions.contextType = ThemeContext;
